\name{cc}
\alias{cc}
\alias{code2HTML}
\alias{code2HTMLjQuery}
\title{cc= code2HTML= code2HTMLjQuery}
\usage{
cc(.theFile = theFile, img = "img", FullSyntaxHighlight = FALSE,
  classicHeaders = FALSE, show = TRUE, toSave = TRUE,
  outSuffix = ".htm")
}
\description{
cc= code2HTML= code2HTMLjQuery
}

