% Generated by roxygen2 (4.0.1): do not edit by hand
\name{code2rmd}
\alias{code2rmd}
\title{code2rmd}
\usage{
code2rmd(.file = get.theFile(), s = readLines(.file, warn = FALSE),
  toTempDir = TRUE, show = TRUE)
}
\description{
similar to knitr::spin
side effect: creates .rmd and .kn.htm  files
}
\examples{
\dontrun{
 rmd= code2rmd()
}
}

