\name{CreateProj}
\alias{CreateNewProj}
\alias{CreateProj}
\alias{CreateProject}
\title{CreateProj= CreateProject= CreateNewProj}
\usage{
CreateProj(newProj.name = "newProjTemplName",
  Templ.dir = "TRUE:/work/UseR-2013/99_commonR/newProjTemplName",
  root = "TRUE:/work/UseR-2013", R2wd = FALSE, overOut = FALSE)
}
\description{
CreateProj= CreateProject= CreateNewProj
}

