\name{createRWJalbum}
\alias{createRWJalbum}
\title{createRWJalbum}
\usage{
RWJournals= MakeRWJournals(...)
createRWJalbum(RWJournals)
}
\arguments{
  \item{RWJournals}{- list of RWJ .html file names}
}
\description{
create RWJ album - galleries for all RWJ in a folder
}
\examples{
\dontrun{
 createRWJalbum(RWJournals, fout='../all.Fig.35.htm')
}
}

