\name{firstFreeFigN}
\alias{firstFreeFigN}
\title{firstFreeFigN}
\usage{
firstFreeFigN(dirr = "../img", patt = "^(Fig|Pic)_(\\\\d+).*\\\\.png$")

firstFreeFigN(dirr = "../img", patt = "^(Fig|Pic)_(\\\\d+).*\\\\.png$")
}
\description{
first Free Fig Number e firstFreeFigN()

firstFreeFigN
}

