\name{gsubInFiles}
\alias{gsubInFiles}
\title{gsubInFiles}
\usage{
gsubInFiles(root = "TRUE:", patt = "\\\\.r\\\\.htm",
  pattRepl = "\\\\.r\\\\.htm", pattOutSub = "\\\\.r\\\\.OO\\\\.htm",
  pattNeg = "^$", exec = FALSE, ...)
}
\description{
sub of multiple patterns in multiple files res<<- is
produced as a side effect for case of error in cycle!!!
}

