% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kn.htm2rwj}
\alias{kn.htm2rwj}
\alias{treat.knit.html}
\title{kn.htm2rwj= treat.knit.html}
\usage{
kn.htm2rwj(theFile, rFile = theFile, dout = dirname(rFile))
}
\description{
kn.htm2rwj= treat.knit.html
}
\examples{
\dontrun{
 treat.knit.html(theFile)  dout=dirname(theFile), rFile=theFile
}
}

