\name{listFuncDef}
\alias{listFuncDef}
\title{listFuncDef}
\usage{
listFuncDef(theFile = "R:/work/R-svn-ass/00_commonR/zBase.r",
  stoplist = "^(c|if|for|function|with|within|exp|log|list|legend|lm|ifelse|invisible|ecdf|lines|plot|min|max|na\\\\.omit|f|expression|options|rgamma|runif|round|str|unique)$")
}
\description{
listFuncDef
}

