\name{listFuncUsage}
\alias{listFuncUsage}
\title{listFuncUsage}
\usage{
listFuncUsage(theFile = "m:/80_ChurnSim/80_ChurnSim.r",
  stoplist = "^(c|if|for|function|with|within|exp|log|list|legend|lm|ifelse|invisible|ecdf|lines|plot|min|max|na\\\\.omit|f|expression|options|rgamma|runif|round|str|unique)$")
}
\description{
listFuncUsage
}

