\name{rwjFig2base64}
\alias{rwjFig2base64}
\title{rwjFig2base64}
\usage{
rwjFig2base64(rwj = "M:/50_HLP/out/HLP_demo/HLP_demo.r.htm")
}
\description{
Insert base64 instead of src ( rwj.htm +/img/) -> rwj.html
- a singlefile
}
\examples{
\dontrun{
 rwjFig2base64(rwj= 'M:/50_HLP/out/HLP_demo/HLP_demo.r.htm')
}
}

