\docType{package}
\name{RWorkJournal}
\alias{RWorkJournal}
\alias{RWorkJournal-package}
\title{R Work Journal and functions for work with code}
\description{
R Work Journal and Helper Funcs for work with code and file
system for a book "Handling Large R Projects"

abbreviations: ds - dataset, patt - pattern, pattNeg -
negative pattern, #e - example
}
\author{
Alex Zolotovitski <alex@zolot.us>
}
\keyword{aliases}

