% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sg}
\alias{saveGraphics}
\alias{sg}
\title{sg= saveGraphics}
\usage{
sg(capt = .main, Width = dev.size(units = "px")[1], off = TRUE,
  Height = dev.size(units = "px")[2], GraphPointSize = 12,
  dirr = "../img", type = "cairo", res = 96, dev = 0,
  fNameWithCapt = FALSE, gg = FALSE, ...)
}
\arguments{
\item{gg}{= is.ggplot}

\item{off}{dev.off() after save}
}
\description{
wrapper for  dev.print  -  save graphics to .png file
}

