% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sgd3}
\alias{sgd3}
\title{sgd3}
\usage{
sgd3(capt = "", dirr = "../img", absPath = TRUE, fNameWithCapt = FALSE,
  fin = ".htm", save = TRUE)
}
\description{
save d3 graphics to jFig_dd.htm file
}
\examples{
\dontrun{
 sgj(p1)
}
}

