\name{subSingleQuote2div}
\alias{subSingleQuote2div}
\title{subSingleQuote2div}
\usage{
subSingleQuote2div(quote = "^\\\\s*`\\\\s*$",
  pattNeg = "'\\\\`'|\\"\\\\`\\"", repl = c("<div class=\\"sq\\">",
  "</div>"), s1 = readLines(theFile, warn = FALSE))
}
\description{
really it used for backticks rather than single quotes :
}

