\name{CreateProj}
\alias{CreateNewProj}
\alias{CreateProj}
\alias{CreateProject}
\title{CreateProj= CreateProject= CreateNewProj}
\usage{
CreateProj(newProj.name = "newProjTemplName",
  Templ.dir = system.file("newProjTemplName", package = "WorkJournal"),
  root = gw(), R2wd = FALSE, overOut = FALSE)
}
\description{
Create New Project from template
}
\examples{
\dontrun{
  CreateNewProj(newProj.name= '96_aaaa',  root='../../')
}
}

