% Generated by roxygen2 (4.0.1): do not edit by hand
\name{firstFreeFigN}
\alias{firstFreeFigN}
\title{firstFreeFigN}
\usage{
firstFreeFigN(dirr = "../img", patt = "^(Fig|Pic)_(\\\\d+).*\\\\.png$")

firstFreeFigN(dirr = "../img", patt = "^(Fig|Pic)_(\\\\d+).*\\\\.png$")
}
\description{
first Free Fig Number

firstFreeFigN
}
\examples{
firstFreeFigN()
}

