\name{sg}
\alias{HHp}
\alias{sg}
\title{sg= HHp}
\usage{
sg(capt = .main, Width = dev.size(units = "px")[1], off = TRUE,
  Height = dev.size(units = "px")[2], GraphPointSize = 12,
  dirr = "../img", type = "cairo", res = 96, dev = 0,
  fNameWithCapt = FALSE, gg = FALSE, ...)
}
\description{
wrapper for dev.print - save graphics to .png file
}

